+++ Auth/src/commonMain/kotlin/io/github/jan/supabase/auth/RedirectUrl.kt
@@ -5,6 +5,7 @@ import io.github.jan.supabase.annotations.SupabaseInternal
 @SupabaseInternal
 internal expect fun Auth.defaultPlatformRedirectUrl(): String?
 
-internal fun Auth.defaultRedirectUrl(): String? {
+@SupabaseInternal
+fun Auth.defaultRedirectUrl(): String? {
     return config.defaultRedirectUrl ?: defaultPlatformRedirectUrl()
 }