+++ Auth/src/commonMain/kotlin/io/github/jan/supabase/auth/mfa/MfaApi.kt
@@ -8,7 +8,6 @@ import io.github.jan.supabase.auth.user.UserMfaFactor
 import io.github.jan.supabase.auth.user.UserSession
 import io.github.jan.supabase.putJsonObject
 import io.github.jan.supabase.safeBody
-import io.ktor.client.call.body
 import io.ktor.util.decodeBase64String
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
@@ -121,14 +120,14 @@ internal class MfaApiImpl(
                 active = current == AuthenticatorAssuranceLevel.AAL2
             )
         } else {
-            MfaStatus(false, false)
+            MfaStatus(enabled = false, active = false)
         }
     }
 
     override val verifiedFactors: List<UserMfaFactor>
         get() = auth.currentUserOrNull()?.factors?.filter(UserMfaFactor::isVerified) ?: emptyList()
 
-    val api = auth.api
+    val api = auth.userApi
 
     override suspend fun <Config, Response> enroll(factorType: FactorType<Config, Response>, friendlyName: String?, config: Config.() -> Unit): MfaFactor<Response> {
         val result = api.postJson("factors", buildJsonObject {
@@ -136,7 +135,7 @@ internal class MfaApiImpl(
             putJsonObject(factorType.encodeConfig(config))
             friendlyName?.let { put("friendly_name", it) }
         })
-        val json = result.body<JsonObject>()
+        val json = result.safeBody<JsonObject>()
         val factorData = factorType.decodeResponse(json)
         return MfaFactor(
             json["id"]!!.jsonPrimitive.content,
@@ -164,7 +163,7 @@ internal class MfaApiImpl(
             put("code", code)
             put("challenge_id", challengeId)
         })
-        val session = result.body<UserSession>()
+        val session = result.safeBody<UserSession>()
         if(saveSession) {
             auth.importSession(session)
         }