+++ Auth/src/tvosMain/kotlin/io/github/jan/supabase/auth/AuthConfig.kt
@@ -1,8 +1,6 @@
 package io.github.jan.supabase.auth
 
-import io.github.jan.supabase.plugins.CustomSerializationConfig
-
 /**
  * The configuration for [Auth]
  */
-actual class AuthConfig : CustomSerializationConfig, AuthConfigDefaults()
+actual class AuthConfig : AuthConfigDefaults()