+++ Supabase/src/commonMain/kotlin/io/github/jan/supabase/PlatformTarget.kt
@@ -12,7 +12,7 @@ enum class PlatformTarget {
 
 data class OSInformation(
     val name: String,
-    val version: String
+    val version: String?
 ) {
 
     companion object {